+++ merlin/models/tf/transforms/sequence.py
@@ -1,0 +1,49 @@
+import tensorflow as tf
+
+from merlin.models.tf.core.base import Block, PredictionOutput
+
+
+@Block.registry.register_with_multiple_names("remove_pad_3d")
+@tf.keras.utils.register_keras_serializable(package="merlin_models")
+class RemovePad3D(Block):
+    """
+    Flatten the sequence of labels and filter out non-targets positions
+
+    Parameters
+    ----------
+        padding_idx: int
+            The padding index value.
+            Defaults to 0.
+    Returns
+    -------
+        targets: tf.Tensor
+            The flattened vector of true targets positions
+        flatten_predictions: tf.Tensor
+            If the predictions are 3-D vectors (sequential task),
+            flatten the predictions vectors to keep only the ones related to target positions.
+    """
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.padding_idx = 0
+
+    def compute_output_shape(self, input_shape):
+        return input_shape
+
+    def call_outputs(
+        self, outputs: PredictionOutput, training=True, **kwargs
+    ) -> "PredictionOutput":
+        targets, predictions = outputs.targets, outputs.predictions
+        targets = tf.reshape(targets, (-1,))
+        non_pad_mask = targets != self.padding_idx
+        targets = tf.boolean_mask(targets, non_pad_mask)
+
+        assert isinstance(predictions, tf.Tensor), "Predictions must be a tensor"
+
+        if len(tuple(predictions.get_shape())) == 3:
+            predictions = tf.reshape(predictions, (-1, predictions.shape[-1]))
+            predictions = tf.boolean_mask(
+                predictions, tf.broadcast_to(tf.expand_dims(non_pad_mask, 1), tf.shape(predictions))
+            )
+
+        return outputs.copy_with_updates(predictions=predictions, targets=targets)
